+++ R/alg_classes.R
@@ -176,8 +176,8 @@ find_model = function(X){
         if(all(round(X)==X)){
           model = methods::new("co_dcsbm")  
         }else{
-          #model = methods::new("gmm",N0=ncol(X)+1,epsilon=100*diag(1/diag(stats::cov(X))),mu=apply(X,2,mean),tau=0.01)
-          model = methods::new("diaggmm",mu=apply(X,2,mean),beta=0.1)
+          model = methods::new("gmm",N0=ncol(X),epsilon=0.1*diag(diag(stats::cov(X))),mu=apply(X,2,mean),tau=0.01)
+          #model = methods::new("diaggmm",mu=apply(X,2,mean),beta=0.1)
         }
       }
     }else{