+++ R/cleanpath.R
@@ -11,7 +11,10 @@ NULL
 
 # extract the pareto front
 extract_front_height=function(sol){
-  
+  if(sol@K==1){
+    H=c(0)
+    return(H)
+  }
   # vector of icls value from root to leaves
   icl=c(sol@icl,sapply(sol@path,function(v){v$icl1}))
   icl = icl[length(icl):1]
@@ -114,8 +117,7 @@ cleanpathopt = function(pathsol){
     
     
     # check for non empty path
-    if(length(path)>0){
-      
+    if(length(pathsol@path)>1){
       # compute the pareto front and extract the height as -log(alpha) of each merge in the front
       Hfront = extract_front_height(pathsol)
       # initialisation